Move flags and other constants from InetSocket onto SceNetInet.
