+++ Core/HLE/sceNetInet.h
@@ -91,6 +91,53 @@ enum PspInetProtocol {
     PSP_NET_INET_IPPROTO_RAW = 255, // raw IP packet
 };
 
+// TODO: INET_
+enum PspInetSocketOptionName {
+    // TODO: also specify minimum socket size
+    INET_SO_ACCEPTCONN   = 0x0002, // socket has had listen()
+    INET_SO_REUSEADDR    = 0x0004, // allow local address reuse
+    INET_SO_KEEPALIVE    = 0x0008, // keep connections alive
+    INET_SO_DONTROUTE    = 0x0010, // just use interface addresses
+    INET_SO_BROADCAST    = 0x0020, // permit sending of broadcast msgs
+    INET_SO_USELOOPBACK  = 0x0040, // bypass hardware when possible
+    INET_SO_LINGER       = 0x0080, // linger on close if data present
+    INET_SO_OOBINLINE    = 0x0100, // leave received OOB data in line
+    INET_SO_REUSEPORT    = 0x0200, // allow local address & port reuse
+    INET_SO_TIMESTAMP    = 0x0400, // timestamp received dgram traffic
+    INET_SO_ONESBCAST    = 0x0800, // allow broadcast to 255.255.255.255
+    INET_SO_SNDBUF       = 0x1001, // send buffer size
+    INET_SO_RCVBUF       = 0x1002, // receive buffer size
+    INET_SO_SNDLOWAT     = 0x1003, // send low-water mark
+    INET_SO_RCVLOWAT     = 0x1004, // receive low-water mark
+    INET_SO_SNDTIMEO     = 0x1005, // send timeout
+    INET_SO_RCVTIMEO     = 0x1006, // receive timeout
+    INET_SO_ERROR        = 0x1007, // get error status and clear
+    INET_SO_TYPE         = 0x1008, // get socket type
+    INET_SO_OVERFLOWED   = 0x1009, // datagrams: return packets dropped
+    INET_SO_NONBLOCK     = 0x1009, // non-blocking I/O
+};
+
+enum PspInetLimit {
+    PSP_NET_INET_SOMAXCONN = 128,
+};
+
+enum PspInetMessageFlag {
+    INET_MSG_OOB = 1,
+    INET_MSG_PEEK = 1 << 1,
+    INET_MSG_DONTROUTE = 1 << 2,
+    INET_MSG_EOR = 1 << 3,
+    INET_MSG_TRUNC = 1 << 4,
+    INET_MSG_CTRUNC = 1 << 5,
+    INET_MSG_WAITALL = 1 << 6,
+    INET_MSG_DONTWAIT = 1 << 7,
+    INET_MSG_BCAST = 1 << 8,
+    INET_MSG_MCAST = 1 << 9
+};
+
+enum InetErrorCode {
+    INET_EINPROGRESS = 119,
+};
+
 class SceNetInet {
 public:
     static bool Init();
@@ -106,13 +153,17 @@ public:
     static bool TranslateInetSocketTypeToNative(int &destSocketType, bool &destNonBlocking, int srcSocketType);
     // TODO: document that errno should be set to EPROTONOSUPPORT when this returns false
     static bool TranslateInetProtocolToNative(int &destProtocol, int srcProtocol);
+    // TODO: document that errno should be set to EPROTONOSUPPORT when this returns false
+    static bool TranslateInetOptnameToNativeOptname(int &destOptname, int inetOptname);
+    static int TranslateInetFlagsToNativeFlags(int messageFlags, bool nonBlocking);
+    static int TranslateNativeErrorToInetError(int nativeError);
 
     // TODO: document
 
     int GetLastError();
     void SetLastError(int error);
     int SetLastErrorToMatchPlatform();
-    std::shared_ptr<InetSocket> CreateAndAssociateInetSocket(int nativeSocketId, bool nonBlocking);
+    std::shared_ptr<InetSocket> CreateAndAssociateInetSocket(int nativeSocketId, int protocol, bool nonBlocking);
     std::shared_ptr<InetSocket> GetInetSocket(int inetSocketId);
     bool GetNativeSocketIdForInetSocketId(int &nativeSocketId, int inetSocketId);
     bool EraseNativeSocket(int inetSocketId);
@@ -127,6 +178,11 @@ private:
     // TODO: document that this does not include flags
     static std::unordered_map<PspInetSocketType, int> gInetSocketTypeToNativeSocketType;
     static std::unordered_map<PspInetProtocol, int> gInetProtocolToNativeProtocol;
+    // TODO: Handle commented out options
+    static std::unordered_map<PspInetSocketOptionName, int> gInetSocketOptnameToNativeOptname;
+    static std::unordered_map<PspInetMessageFlag, int> gInetMessageFlagToNativeMessageFlag;
+
+    static std::unordered_map<int, InetErrorCode> gNativeErrorCodeToInetErrorCode;
 
     int mLastError = 0;
     std::unordered_map<int, std::shared_ptr<InetSocket>> mInetSocketIdToNativeSocket;
